% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargaRDS.R
\name{cargaRDS}
\alias{cargaRDS}
\title{cargar rx RDS}
\usage{
cargaRDS(rxdir)
}
\arguments{
\item{rxdir}{El directorio de donde estan los jpg de rx}
}
\value{
Un data frame con la informacion del RDS y el folder
}
\description{
Carga el folde de RX y crea un archivo RDS con todos los datos, si no existe lo crea.
}
\examples{
# library(stringr); library(dplyr)
# test <- cargaRDS("D:/OneDrive/BarbiApp/AppShiny/testFolder")
}
